refactor: change modules import to use relative paths (#14)
* relative paths for components till GiftForm.tsx

* refactor: change modules import to use relative paths

* save commit
