+++ ui-react/src/components/index.ts
@@ -1,3 +1,24 @@
-import { GiftEdit } from './gift';
+import { GiftEdit, GiftAdd } from './gift';
+import { CategoriesList } from './categories-list';
+import { CategoryEdit, CategoryAdd } from './category';
+import { GiftedUserEdit, GiftedUserAdd } from './gifted-user';
+import { GiftedUserList } from './gifted-user-list';
+import { GiftsList } from './gifts-list';
+import { Layout, AdminLayout } from './layout';
 
-export { GiftEdit };
+import { TranslatedText } from './translated-text';
+
+export {
+  GiftEdit,
+  GiftAdd,
+  TranslatedText,
+  CategoryEdit,
+  GiftedUserEdit,
+  GiftsList,
+  Layout,
+  AdminLayout,
+  CategoryAdd,
+  GiftedUserAdd,
+  GiftedUserList,
+  CategoriesList,
+};
