+++ ui-react/src/pages/admin-panel/user-gifts-management/GiftsPerSelectedUser.tsx
@@ -6,17 +6,12 @@ import { SelectChangeEvent, Typography } from '@mui/material';
 
 import { useNavigate, useParams } from 'react-router-dom';
 
-import { GiftsList } from 'components/gifts-list/GiftsList';
-import {
-  SelectField,
-  SelectFieldOption,
-} from 'components/material/SelectField';
+import { GiftsList } from 'components';
+import { SelectField, SelectFieldOption } from 'components/material';
 
 import { TranslatedText } from 'components/translated-text/TranslatedText';
 
-import { useGiftedUsers } from '../../../api/hooks/gifted-users';
-
-import { GiftsQueryKeys } from '../../../api/hooks/gifts';
+import { useGiftedUsers, GiftsQueryKeys } from 'api/hooks';
 
 export const GiftsPerSelectedUser = () => {
   const [giftedUsersDropdownOptions, setGiftedUsersDropdownOptions] = useState<
